imported data from Sqlite3 to MySql

+++ src/main/java/com/nks/quotejava2/repositories/mysql/InfoRepository.java
@@ -3,7 +3,5 @@ package com.nks.quotejava2.repositories.mysql;
 import com.nks.quotejava2.models.mysql.Info;
 import org.springframework.data.repository.CrudRepository;
 
-import java.math.BigInteger;
-
-public interface InfoRepository extends CrudRepository<Info, BigInteger> {
+public interface InfoRepository extends CrudRepository<Info, Long> {
 }
